Resta de gastos a saldo

+++ app/src/main/java/com/itesm/budget/ui/Gastos/RegistroGasto.kt
@@ -2,5 +2,5 @@ package com.itesm.budget.ui.Gastos
 
 data class RegistroGasto(var compra : String = "",
                          var categoria: String="",
-                         var gasto: Double = 0.0,
+                         var gasto: Float = 0.0f,
                          var fecha: String = "")
